Update schema rules to throw readable errors
When the developer attaches a schema to a collection, the resulting
deny rule with throw an error if schema validation fails, this provides
feedback to the user on why the update or insert failed.

Note that CollectionRule.allowUpdate (etc.) do not throw errors and do
not provide any user feedback as to why the db operation failed but
return a generic ‘403’ error.